Split the logic between ZoneMap pure component and Map layout wrapper with layers
